/editorに<Canvas>を移動
