+++ pages/editor/index.tsx
@@ -1,36 +1,44 @@
-import React, { useRef, useState } from 'react'
-import { Canvas, useFrame } from '@react-three/fiber'
-import * as THREE from 'three/src/Three';
+import { NextPage } from 'next';
+import { useRef, useState } from 'react';
+import { Canvas, useFrame } from '@react-three/fiber';
+import { DstAlphaFactor, Mesh } from 'three';
 
-function Box(props) {
-  // This reference will give us direct access to the mesh
-  const mesh = useRef()
-  // Set up state for the hovered and active state
-  const [hovered, setHover] = useState(false)
-  const [active, setActive] = useState(false)
-  // Subscribe this component to the render-loop, rotate the mesh every frame
-  useFrame((state, delta) => (mesh.current.rotation.x += delta))
-  // Return view, these are regular three.js elements expressed in JSX
+type BoxProps = {
+  position: [x: number, y: number, z: number];
+};
+
+const Box: React.FC<BoxProps> = (props) => {
+  const mesh = useRef<Mesh>(null!);
+  const [hovered, setHover] = useState(false);
+  const [active, setActive] = useState(false);
+  useFrame(() => (mesh.current.rotation.x += 0.01));
+
   return (
     <mesh
       {...props}
       ref={mesh}
       scale={active ? 1.5 : 1}
-      onClick={(event) => setActive(!active)}
-      onPointerOver={(event) => setHover(true)}
-      onPointerOut={(event) => setHover(false)}>
+      onClick={() => setActive(!active)}
+      onPointerOver={() => setHover(true)}
+      onPointerOut={() => setHover(false)}
+    >
       <boxGeometry args={[1, 1, 1]} />
       <meshStandardMaterial color={hovered ? 'hotpink' : 'orange'} />
     </mesh>
-  )
-}
-export const R3f = () => {
-   return(
+  );
+};
+
+const Editor = () => {
+  return(
+    <div style={{ width: '100vw', height: '100vh' }}>
     <Canvas>
-    <ambientLight />
-    <pointLight position={[10, 10, 10]} />
-    <Box position={[-1.2, 0, 0]} />
-    <Box position={[1.2, 0, 0]} />
+      <ambientLight />
+      <pointLight position={[10, 10, 10]} />
+      <Box position={[-1.2, 0, 0]} />
+      <Box position={[1.2, 0, 0]} />
     </Canvas>
-    ) 
+  </div>
+  )
+  
 }
+export default Editor;
